+++ src/main/java/com/oo2/grupo17/dtos/records/RegisterClienteDto.java
@@ -12,7 +12,7 @@ public record RegisterClienteDto (
 		@NotBlank String password,
 		@NotBlank String nombre,
 		@NotNull @Min(10000000) @Max(99999999) int dni,
-		@NotNull @Min(1000000000) @Max(9999999999L) int movil,
+		@NotNull @Min(1000000000) @Max(9999999999L) long movil,
 		@Min(1000000) @Max(99999999) int telefono
 		
 ) {}
